refactoring the way to download a photo --> static method added in FireBaseCloud to handle it
