+++ app/src/main/res/values/strings.xml
@@ -48,6 +48,6 @@
     <string name="initial_greeting_1">Welcome, Jean-Louis.</string>
     <string name="initial_greeting_2">Your Sciper number is 123456.</string>
     <string name="problem_images_directory_name">problem_images</string>
-    <string name="user_pictures_directory_name">user_pictures</string>
+    <string name="FB_user_pictures_directory_name">user_pictures</string>
 
 </resources>
